Using profile image and replacing blank users
