+++ src/components/Buttons/UserButton.scss
@@ -6,8 +6,8 @@
     width: 50px;
     height: 50px;
     border-radius: 50%;
-    background-color: blueviolet;
     margin-right: 10px;
+    object-fit: cover;
   }
 
   p {
