+++ src/components/Nav/Nav.tsx
@@ -102,7 +102,7 @@ function Nav() {
           onClick={() => handleClick('profile')}
         >
           {showNavLi ? 
-            <UserButton className={selected === 'profile' ? 'selected' : ''}  hasInfo userInfo={{name: userNameAndExp?.fullname, desc: userNameAndExp?.investExp}}/>
+            <UserButton className={selected === 'profile' ? 'selected' : ''}  hasInfo userInfo={{name: userNameAndExp?.fullname, desc: userNameAndExp?.investExp}} userImg={userNameAndExp?.profileImg}/>
             :
             <li className='nav_ul_a_li'>
               <img className="user-img"/>
